+++ CHANGES.md
@@ -4,3 +4,5 @@
 * can describe SPNs in transport layer in real-time (as their bytes are received) (thanks @j4l)
 * correctly reassembles  RTS-CTS transport sessions as well (thanks @j4l)
 * can specify J1939 JSON db on command line (thanks @j4l)
+* default to describing transport PGNs as first-class PGN
+* default to omitting description of incomplete frames